refactored permutation mutation operator tests
